feat: Add commitlint to lint commit messages
Create a commit convention for our organization, install `commitlint`
and configure it as a hook to lint commit message. This helps will your
team and contribuitors to adhering to the said commit convention.

Add a `commitlint` script in Travis to catch anyone that tries to bypass
the local validation.

Update `CONTRIBUTING.md` to reflect the convention created by this
commit.

Closes #23