chore: Add issue templates
Without templates, sometimes the issue author can forget to add essential information to the resolution, not only that but issues can also be described in different words and formats each time, which makes them more difficult to standardize.

It’s hard to solve a problem when important details are missing and every issue is diferrent. This change add issue templates to help contributors share the right details at the start of a thread and keep things consistent.