chore: Add tests and remove default validation
As the new API includes a lot of new methods we needed to add the
remaining tests to Jest.

Also, now Unform does not provide built in validation with Yup, instead,
the validation must be made by the final user inside the submit method
and, after that, the `setErrors` method can be used to provide all input
errors.