chore: Add support for objects/arrays in setData
Currently if the `setData` function was called with an object as input
value, the value was parsed as new form fields.

Example:
setData({ tech: { id: 5, label: 'React' } })

Unform would try to find inputs name tech.id and tech.label even if they
didn't exists.

Now, unform parses the setData value according to the inputs that were
registered, so, in the example, it will set "tech" input value as { id:
5, label: 'React' }